recycler view is ready. need 2 correct bacground color for night time & make title, & make screen gotation both

+++ src/main/java/com/example/racertimer/forecast/data/parsers/ParserJsonToQueueLines.kt
@@ -20,7 +20,7 @@ class ParserJsonToQueueLines {
             val windJson = currentJson.getJSONObject("wind")
             val windSpeed = windJson.getInt("speed")
             val windDir = windJson.getInt("deg")
-            val windGust = windJson.getDouble("gust")
+            val windGust = windJson.getInt("gust")
 
             val currentLine = ForecastLine(
                 time = time,
